# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Evamed Connector API
  description: |
    OpenAPI v3 specification exposing read-only operations for EVAMED data.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Projects
    description: Endpoints related to projects listing and reference data.
  - name: Global-Indicators
    description: Aggregated indicators and statistics.
paths:
  /projects:
    get:
      summary: List projects
      description: |
        Returns a paginated list of projects. Optional filters for project manager and
        management assistant can be combined with logical AND. Results are ordered by
        creation date descending unless a custom sort is provided.
      operationId: getProjects
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
        - name: projectManagerId
          in: query
          required: false
          description: Filter by the identifiers of project managers assigned to the project. Accepts multiple values.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: managementAssistantId
          in: query
          required: false
          description: Filter by the identifiers of management assistants assigned to the project. Accepts multiple values.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: projectType
          in: query
          required: false
          description: Filter by project types. Accepts multiple values.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: isSensitive
          in: query
          required: false
          description: Filter by sensitive flag.
          schema:
            type: boolean
        - name : isProjectFinished
          in: query
          required: false
          description: Filter by project finished flag.
          schema: 
            type: boolean
        - name: phase
          in: query
          required: false
          description: |
            Project phase filter possible values :
            - saisine
            - debut_cadrage
            - cadrage
            - pgt
            - dgt
            - examen
            - validation
            - mise_en_ligne
            - cloture
          schema:
            $ref: '#/components/schemas/ProjectPhaseFilter'

        - name: period
          in: query
          required: false
          description: |
            Period filter containing a start date and an end date (inclusive).
            Uses deepObject serialization: period[beginDate]=YYYY-MM-DD&period[endDate]=YYYY-MM-DD.
          style: deepObject
          explode: true
          schema:
            $ref: '#/components/schemas/PeriodFilter'
      responses:
        "200":
          description: Paginated list of projects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectResponse'
              examples:
                sample:
                  summary: Example project list
                  value:
                    items:
                      - id: "df6d2a95-3b7e-4f4f-8d4b-31d2a72bfb1e"
                        number: "PRJ-2024-001"
                        type: "clinical"
                        title: "Study on treatment efficacy"
                        isSensitive: false
                        applicationDomains:
                          - "pharmaceutical"
                        phases:
                          - beginDate: "2024-01-15"
                            endDate: "2024-06-15"
                            label: "Phase I"
                        projectManagers:
                          - firstName: "Alex"
                            lastName: "Durand"
                            isPrimary: true
                            isSecondary: false
                        managementAssistant: "Louise Martin"
                        creationAt: "2024-01-01T10:00:00Z"
                        endAt: "2024-12-31T18:00:00Z"
                    limit: 20
                    offset: 0
                    total: 1
        "400":
          description: Invalid query parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No projects matched the provided filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/managers:
    get:
      summary: List project managers
      description: Returns project managers filtered by user type.
      operationId: getProjectManagers
      tags:
        - Projects
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: List of project managers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectManager'
        "400":
          description: Invalid query parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/management-assistants:
    get:
      summary: List management assistants
      description: Returns management assistants filtered by user type.
      operationId: getManagementAssistants
      tags:
        - Projects
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: List of management assistants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementAssistant'
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/typologies:
    get:
      summary: List project typologies
      description: Returns available typologies filtered by user type.
      operationId: getTypologies
      tags:
        - Projects
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: List of typologies.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Invalid query parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /global-indicators/projects-by-typology:
    get:
      summary: Count of projects by typology
      description: Returns the number of projects grouped by typology.
      operationId: getProjectsByTypology
      tags:
        - Global-Indicators
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: Map of typology identifiers to project counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringIntegerMap'
              example:
                clinical: 12
                epidemiological: 7
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /global-indicators/projects-by-phase:
    get:
      summary: Count of projects by phase
      description: Returns the number of projects grouped by current phase.
      operationId: getProjectsByPhase
      tags:
        - Global-Indicators
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: Map of project phases to counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringIntegerMap'
              example:
                cadrage: 5
                examen: 3
                mise_en_ligne: 2
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /global-indicators/average-time-per-typology:
    get:
      summary: Average time by typology
      description: Returns the average processing time for projects by typology.
      operationId: getAverageTimePerTypology
      tags:
        - Global-Indicators
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: Map of typology identifiers to average time values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringIntegerMap'
              example:
                clinical: 180
                epidemiological: 95
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /global-indicators/working-group-for-current-year:
    get:
      summary: Working group for the current year
      description: Returns the working group identifier for the current year.
      operationId: getWorkingGroupForCurrentYear
      tags:
        - Global-Indicators
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: Working group identifier for the current year.
          content:
            application/json:
              schema:
                type: string
              example: "working-group-2024"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /global-indicators/project-typology-per-domain:
    get:
      summary: Typology distribution per domain
      description: Returns the number of projects by typology within each application domain.
      operationId: getProjectTypologyPerDomain
      tags:
        - Global-Indicators
      parameters:
        - name: userType
          in: query
          required: true
          description: Type of user issuing the request. Value must be "sbp" for SBP Service Manager or "sr" for SR Service Manager.
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        "200":
          description: Number of projects by typology within each application domain.
          content:
            application/json:
              schema:
                type: array
                description: List of typology distribution per domain.
                items:
                  $ref: '#/components/schemas/TypologyPerDomainGlobalIndicatorResponse'
              example:
                pharmaceutical:
                  clinical: 4
                  epidemiological: 2
                medical_device:
                  clinical: 3
                  epidemiological: 1
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UserType:
      type: string
      description: Type of user issuing the request.
      enum: [sbp, sr]
      x-enum-varnames:
        - SBP_SERVICE_MANAGER
        - SR_SERVICE_MANAGER
      x-enum-descriptions:
        - SBP Service Manager
        - SR Service Manager
    StringIntegerMap:
      type: object
      description: Map of string keys to integer values.
      additionalProperties:
        type: integer
        format: int32
    ProjectPhaseFilter:
      type: string
      description: Project phase.
      enum:
        - SAISINE
        - DEBUT_CADRAGE
        - CADRAGE
        - PGT
        - DGT
        - EXAMEN
        - VALIDATION
        - MISE_EN_LIGNE
        - CLOTURE
    PeriodFilter:
      type: object
      description: Objet période utilisé comme filtre de recherche, avec une date de début et une date de fin.
      properties:
        beginDate:
          type: string
          format: date
          description: Date de début (incluse)
        endDate:
          type: string
          format: date
          description: Date de fin (incluse)
      example:
        beginDate: '2024-01-01'
        endDate: '2024-12-31'
    PaginatedProjectResponse:
      type: object
      description: Standard paginated response envelope for projects.
      required:
        - items
        - limit
        - offset
        - total
      properties:
        items:
          type: array
          description: List of projects in the current page.
          items:
            $ref: '#/components/schemas/Project'
        limit:
          type: integer
          description: Number of items requested per page.
          minimum: 1
          maximum: 100
          default: 20
        offset:
          type: integer
          description: Zero-based index of the first item returned.
          minimum: 0
          default: 0
        total:
          type: integer
          description: Total number of projects matching the filters (without pagination).
          minimum: 0
    Project:
      type: object
      required:
        - id
        - number
        - type
        - title
        - isSensitive
        - applicationDomains
        - phases
        - projectManagers
        - managementAssistant
        - creationAt
      properties:
        id:
          type: string
          description: Unique identifier of the project.
        number:
          type: string
          description: Business reference number of the project.
        typeCode:
          type: string
          description: Project type code.
        typologies:
          type: array
          items:
            type: string
        title:
          type: string
          description: Descriptive title of the project.
        isSensitive:
          type: boolean
          description: Indicates whether the project handles sensitive data.
        applicationDomains:
          type: array
          items:
            type: string
          description: Application domains of the project.
        projectManagers:
          type: object
          description: Grouped lists of project managers.
          properties:
            primary:
              type: array
              description: List of primary managers supervising the project.
              items:
                $ref: '#/components/schemas/ProjectManager'
            secondary:
              type: array
              description: List of secondary managers supervising the project.
              items:
                $ref: '#/components/schemas/ProjectManager'
        managementAssistant:
          type: array
          description: Assistants assigned to the project management team.
          items:
            $ref: '#/components/schemas/ManagementAssistant'
        referralDate:
          type: string
          format: date
          description: Date of referral (Date de saisine).
        scopingStartDate:
          type: string
          format: date
          description: Date of starting scoping (Date début de cadrage)
        scopingDate:
          type: string
          format: date
          description: Date of scoping (Date de cadrage).
        pgtDate:
          type: string
          format: date
          description: Date of PGT.
        dgtDate:
          type: string
          format: date
          description: Date of DGT.
        reviewDate:
          type: string
          format: date
          description: Date of review (Date d'examen).
        validationDate:
          type: string
          format: date
          description: Date of validation (Date de validation).
        publicationDate:
          type: string
          format: date
          description: Date of publication (Date de publication).
        closureDate:
          type: string
          format: date
          description: Date of closure (Date de clôture).
        plannedScopingStartDate:
          type: string
          format: date
          description: Date of starting scoping (Date début de cadrage)
        plannedScopingDate:
          type: string
          format: date
          description: Date of scoping (Date de cadrage).
        plannedPgtDate:
          type: string
          format: date
          description: Date of PGT.
        plannedDgtDate:
          type: string
          format: date
          description: Date of DGT.
        plannedReviewDate:
          type: string
          format: date
          description: Date of review (Date d'examen).
        plannedValidationDate:
          type: string
          format: date
          description: Date of validation (Date de validation).
        plannedPublicationDate:
          type: string
          format: date
          description: Date of publication (Date de publication).
        plannedClosureDate:
          type: string
          format: date
          description: Date of closure (Date de clôture).
        gtglMeetings:
          type: array
          description: List of GTGL meetings held during the project.
          items:
            $ref: '#/components/schemas/GtGlMeeting'
        commissionSessions:
          type: array
          description: List of commission sessions held during the project.
          items:
            $ref: '#/components/schemas/CommissionSession'
        suspensions:
          type: array
          description: List of suspensions applied to the project.
          items:
            $ref: '#/components/schemas/Suspension'
    ProjectManager:
      type: object
      required:
        - id
        - firstName
        - lastName
      properties:
        id:
          type: string
          description: Unique identifier of the project manager.
        firstName:
          type: string
          description: Manager's given name.
        lastName:
          type: string
          description: Manager's family name.
    ManagementAssistant:
      type: object
      required:
        - id
        - firstName
        - lastName
      properties:
        id:
          type: string
          description: Unique identifier of the management assistant.
        firstName:
          type: string
          description: Management assistant's given name.
        lastName:
          type: string
          description: Management assistant's family name.
    GtGlMeeting:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Date of the meeting.
        label:
          type: string
          description: Label of the meeting.
    CommissionSession:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Date of the session.
        label:
          type: string
    Suspension:
      type: object
      properties:
        startDate:
          type: string
          format: date
          description: Start date of the suspension.
        endDate:
          type: string
          format: date
          description: End date of the suspension.
        label:
          type: string
          description: Label of the suspension.
        comment:
          type: object
          description: Comment associated with the suspension.
    TypologyPerDomainGlobalIndicatorResponse:
      type: object
      description: Response envelope for global indicators based on project typology.
      required:
        - domainLabel
        - typology
        - count
      properties:
        domainLabel:
          type: string
          description: Application domain label.
        typology:
          type: string
          description: Project typology.
        count:
          type: integer
          description: Number of projects with the given typology in the given domain.
    ErrorResponse:
      type: object
      description: Standard error envelope returned by the API.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Application-specific error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          description: Optional additional information about the error.
          nullable: true
          oneOf:
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true
  parameters:
    Offset:
      name: offset
      in: query
      description: Zero-based index of the first item to return.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: Number of items to return (maximum 100).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20